marrtino_chatbot/requirements.txt
Flask
openai
python-aiml
requests
gtts
websockets
vosk
sounddevice
